I Lied... New cf_relaign_encoders() method in CF_Hardware
This new method will try to see if one encoder has gone farther than
the other, and will try to sync them back up.  I don’t know if we’ll
ever have to use this method, but I figured it would be beneficial to
have just in case we ever have to use it.