Changes as of 2/11/16
So in this I cleaned up some auto code and I added support for a touch
sensor which will hopefully be added onto the climber arm and will stop
the arm before it tries to burn up the servo.  Also, I made the
CF_Hardware and CF_HardwareRed files better by using variable names for
BucketServoPos and ZipLineServoPos.  So that’s about all.  Hopefully,
I’ll soon add support for an ultrasonic distance sensor, because it
looks like the optical distance sensor only works up to about 8 cm,
which is useless for my application.  Also, I plan to modify the path
the robot takes during auto mode to get it over in time to drop the
climbers.